slot: move storage operation types into slot modules
