+++ crates/cranpose-core/src/slot/table/values.rs
@@ -1,5 +1,5 @@
 use super::SlotTable;
-use crate::slot_storage::ValueSlotId;
+use crate::slot::ValueSlotId;
 
 impl SlotTable {
     fn checked_value_slot(&self, slot: ValueSlotId) -> (usize, usize) {
