+++ crates/cranpose-core/src/slot/validate/anchors.rs
@@ -1,6 +1,6 @@
 use super::super::{AnchorState, GroupRecord, SlotTable};
 use super::SlotInvariantError;
-use crate::{collections::map::HashSet, slot_storage::GroupKey, AnchorId};
+use crate::{collections::map::HashSet, slot::GroupKey, AnchorId};
 
 pub(super) fn validate_active_group_anchor(
     table: &SlotTable,
