+++ crates/cranpose-core/src/slot/writer/frames.rs
@@ -1,7 +1,7 @@
 use super::siblings::SiblingIndex;
 use crate::{
     collections::map::{HashMap, HashSet},
-    slot_storage::GroupKey,
+    slot::GroupKey,
     AnchorId, Key,
 };
