+++ crates/cranpose-core/src/slot/writer/nodes.rs
@@ -1,7 +1,7 @@
 use super::super::{
     collect_root_node_ids_from_records_into, NodeRecord, SlotTable, SlotWriteSession,
 };
-use crate::{slot_storage::NodeSlotUpdate, AnchorId, NodeId};
+use crate::{slot::NodeSlotUpdate, AnchorId, NodeId};
 
 impl SlotTable {
     fn collect_subtree_node_records(&self, group_anchor: AnchorId) -> Vec<NodeRecord> {
