+++ crates/cranpose-core/src/slot/writer/payload.rs
@@ -1,5 +1,5 @@
 use super::super::{PayloadKind, SlotWriteSession};
-use crate::{slot_storage::ValueSlotId, Owned};
+use crate::{slot::ValueSlotId, Owned};
 
 impl SlotWriteSession<'_> {
     pub(crate) fn value_slot_with_kind<T: 'static>(
